Fix SmartEnum circular dependency problem
e10fe030dd312148837160bc3d12c57e677846e2,
8ad37beebc873220f2eb29f1fa31b96db61a3b48 and other commits were written
to work around a flaw in our implementation of SmartEnum's STI
functionality.  This commit fixes that issue.

Background:

Any code that runs in a rails app must play nice with the rails
autoloader.  One important consequence of using the autoloader is that
class definition order is nondeterministic.  Constants will be loaded in
either the order that they happen to be referenced at runtime (in
development) or by the pseudo-alphabetical order used by
Rails.application.eager_load! (in production).

The autoloader's nondeterminism poses a problem for SmartEnum's STI
functionality.  Consider an application with the following files:

    # my_sti_parent.rb
    class MyStiParent < SmartEnum
      attribute :id, Integer
      register_values_from_file!
    end

    # my_sti_child.rb
    class MyStiChild < MyStiParent
    end

    # data/lookups/my_sti_parents.yml
    ---
    - id: 1
      type: MyStiParent
    - id: 2
      type: MyStiChild

In a new rails process where neither class has yet been loaded, if
MyStiChild is referenced or the eager_load process chooses to require
my_sti_child.rb first, the autoloader will read my_sti_child.rb and
encounter the unloaded constant MyStiParent, which will cause it to
require my_sti_parent.rb.  One of the things it has to do while loading
my_sti_parent.rb is `register_values_from_file!`, which will eventually
result in `MyStiChild.new(id: 2)` being executed.  Since the MyStiChild
constant is still in the process of being loaded we don't have access to
the constant; its use represents a circular dependency that the
autoloader will prevent.

There are a number of ways of resolving this problem using an approach
where SmartEnum instances are only constructed after all classes have
been loaded.  The problem with most of these is that they break code
reloading in development mode.

In order to keep the ergonomics nice for rails console `reload!` and
rails server automatic reloading, I've implemented an alternate
approach whose only downside is increased complexity.  When a base class
calls register_values in a way that involves STI children, we do not
immediately attempt to constantize the type discriminator column in our
attribute hashes.  Instead we load the raw hashes into an intermediate
array whose contents will be instantiated when the enum is locked.  We
then spawn a new thread to lock the enum after waiting long enough to
ensure that ruby is no longer processing the segment of the dependency
tree that would have caused a circular dependency error.

While the second thread is sleeping there is a risk that other code will
access the enum, so we wrap all read access in a mutex that stays locked
for the entire time the second thread is sleeping.  This means that the
200ms interval won't block other code unless some code tries to access
an enum that is currently waiting to process deferred attribute hashes.