Don't use Enumerable feature not in ruby 2.4
Ruby 2.5 added case equality checking to enumerable methods like all?
when no block is passed, similar to Enumerable#grep.  I forgot that this
functionality isn't present in previous versions.

When replacing it with a more compatible equivalent, I benchmarked the
various ways of checking that all elements are symbols, here are the
results which indicate that the case equality operator is marginally
fastest:

    $arr = ('A'..'z').map(&:to_sym)
    Benchmark.ips do |x|
      x.report("is_a?") { $arr.all?{|el| el.is_a?(Symbol) } }
      x.report("instance_of?") { $arr.all?{|el| el.instance_of?(Symbol) } }
      x.report("class ==") { $arr.all?{|el| el.class == Symbol } }
      x.report("===") { $arr.all?{|el| Symbol === el } }
      x.compare!
    end

    Comparison:
                 ===:   247611.2 i/s
        instance_of?:   244526.0 i/s - same-ish: difference falls within error
               is_a?:   238475.1 i/s - same-ish: difference falls within error
            class ==:   214453.4 i/s - 1.15x  slower