Split up test suite into two pieces to avoid AS contamination
This separates spec/active_record_compatibility_spec.rb into its own
individual suite.  The reason for this is that that test needs to
require activerecord which also requires activesupport.  That
in turn loads a number of monkeypatches that we don't want our code to
depend on, but the way rspec works means that mixing the tests will
cause activesupport to be loaded for everything.  To fix this, isolate
the require into a separate test suite that runs last.