Add when_nil option to belongs_to_enum macro
Use the `when_nil` option to specify the ID of a SmartEnum value that should be
returned when the foreign_key reference is NULL.

Helps to to implement the Null Object pattern, or provide a sane default.

It does not change what is written to the association, just what is returned.